Separated the interface of buffered out from implementation
