+++ spinn_front_end_common/interface/interface_functions/front_end_common_buffer_manager_creater.py
@@ -21,24 +21,16 @@ class FrontEndCommonBufferManagerCreater(object):
             placements, tags, txrx, reports_states, app_data_folder)
 
         for placement in placements.placements:
-            if isinstance(placement.subvertex,
-                          AbstractSendsBuffersFromHostPartitionedVertex):
-
+            if isinstance(
+                    placement.subvertex,
+                    AbstractSendsBuffersFromHostPartitionedVertex):
                 if placement.subvertex.buffering_input():
-
-                    # Add the vertex to the managed vertices
                     buffer_manager.add_sender_vertex(placement.subvertex)
 
             if isinstance(placement.subvertex, AbstractReceiveBuffersToHost):
-                    if placement.subvertex.buffering_output:
-                        buffer_manager.add_receiving_vertex(
-                            placement.subvertex)
-
-            # Partitioned vertices can also be output buffered
-            if isinstance(placement.subvertex, AbstractReceiveBuffersToHost):
-                if placement.subvertex.buffering_output:
+                if placement.subvertex.buffering_output():
                     buffer_manager.add_receiving_vertex(placement.subvertex)
-                    placement.subvertex.buffer_manager = buffer_manager
+
             progress_bar.update()
         progress_bar.end()
